+++ packages/react-querybuilder/src/index.debug.ts
@@ -1,0 +1,4 @@
+export * from './barrel';
+export * from './components/index.debug';
+export { QueryBuilder as default } from './components/index.debug';
+export * from './redux/index.debug';